tests/wpt/tests/html/canvas/offscreen/canvas-context/2d.canvas.context.type.exists.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.canvas.context.type.exists</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.context.type.exists</h1>
<p class="desc">The 2D context interface is a property of 'window'</p>

<p class="notes">Defined in "Web IDL" (draft)
<script>
var t = async_test("The 2D context interface is a property of 'window'");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assert(window.OffscreenCanvasRenderingContext2D, "window.OffscreenCanvasRenderingContext2D");
  t.done();

});
</script>
